+++ src/main/java/com/ufc/br/mvp/service/RecebedorService.java
@@ -18,7 +18,7 @@ public class RecebedorService {
 		return repository.save(recebedor);
 	}
 	
-	public void delete(Long id) {
+	public void delete(Integer id) {
 		try {
 			repository.deleteById(id);
 		} catch (IllegalArgumentException e) {
@@ -26,7 +26,7 @@ public class RecebedorService {
 		}
 	}
 	
-	public Recebedor find(Long id) {
+	public Recebedor find(Integer id) {
 		return repository.getOne(id);
 	}
